requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
google-generativeai==0.3.1
requests==2.31.0
python-multipart==0.0.6